// 202401EffectiveAngular/business-tools-monorepo/libs/finance/ui/expenses-registration-forms/src/lib/data-models/dynamic-form.interface.ts
import {ValidatorFn} from "@angular/forms";

export interface DynamicControl {
  controlKey: string;
  formFieldType?: 'input' | 'select';
  inputType?: string;
  label?: string;
  defaultValue?: any;
  selectOptions?: string[];
  updateOn: 'change' | 'blur' | 'submit';
  validators?: ValidatorFn[];
}